Reorganized code according to CR - naming for ComponentType enum has been updated - separated static MainSite footer into a separate file
